+++ app/src/main/java/remix/myplayer/bean/mp3/Artist.kt
@@ -4,6 +4,6 @@ package remix.myplayer.bean.mp3
  * Created by Remix on 2017/10/22.
  */
 
-data class Artist(val artistID: Int,
+data class Artist(val artistID: Long,
                   val artist: String,
                   val count: Int)
